+++ CHANGELOG.md
@@ -4,9 +4,18 @@
 
 ## [0.7.0](https://github.com/Blobfolio/write_atomic/releases/tag/v0.7.0) - TBD
 
+### New
+
+* Re-export `tempfile`
+* Re-export `filetime`
+
 ### Changed
 
 * Bump MSRV to `1.87`
+* `copy_file` now (tries to) copy access and modification times too
+* Improved documentation
+* Improved test coverage
+* Miscellaneous code cleanup and lints